+++ src/main/java/io/coodoo/framework/jpa/boundary/entity/BaseEntity.java
@@ -35,7 +35,7 @@ import javax.persistence.MappedSuperclass;
  */
 @SuppressWarnings("serial")
 @MappedSuperclass
-public abstract class AbstractEntity implements Serializable {
+public abstract class BaseEntity implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -69,7 +69,7 @@ public abstract class AbstractEntity implements Serializable {
         if (this == obj) {
             return true;
         }
-        if (obj == null || !(obj instanceof AbstractEntity)) {
+        if (obj == null || !(obj instanceof BaseEntity)) {
             return false;
         }
 
@@ -77,10 +77,10 @@ public abstract class AbstractEntity implements Serializable {
             return false;
         }
 
-        if (((AbstractEntity) obj).getId() == null) {
+        if (((BaseEntity) obj).getId() == null) {
             return false;
         }
-        return ((AbstractEntity) obj).getId().equals(getId());
+        return ((BaseEntity) obj).getId().equals(getId());
     }
 
 }