+++ src/main/java/io/coodoo/framework/jpa/boundary/entity/RevisionDatesDmEntity.java
@@ -6,16 +6,15 @@ import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
 
 /**
- * Base entity providing identification, optimistic concurrency control and automatically sets creation/update/deletion timestamps
+ * Base entity providing identification and automatically sets creation/update/deletion timestamps
  * 
  * <br>
  * <br>
- *
+ * 
  * <table border="1" summary="Fields">
  * <tr>
  * <th>Fields</th>
  * <th>ID</th>
- * <th>OCC</th>
  * <th>Creation Date</th>
  * <th>Update Date</th>
  * <th>Deletion Date</th>
@@ -23,7 +22,6 @@ import javax.persistence.MappedSuperclass;
  * <tr>
  * <td>Name</td>
  * <td>id</td>
- * <td>version</td>
  * <td>createdAt</td>
  * <td>updatedAt</td>
  * <td>deletedAt</td>
@@ -31,7 +29,6 @@ import javax.persistence.MappedSuperclass;
  * <tr>
  * <td>Type</td>
  * <td>Long</td>
- * <td>Integer</td>
  * <td>LocalDateTime</td>
  * <td>LocalDateTime</td>
  * <td>LocalDateTime</td>
@@ -39,7 +36,6 @@ import javax.persistence.MappedSuperclass;
  * <tr>
  * <td>Column name</td>
  * <td>id</td>
- * <td>version</td>
  * <td>created_at</td>
  * <td>updated_at</td>
  * <td>deleted_at</td>
@@ -50,7 +46,7 @@ import javax.persistence.MappedSuperclass;
  */
 @SuppressWarnings("serial")
 @MappedSuperclass
-public abstract class AbstractVersionRevisionDatesDeleteMarkerEntity extends AbstractVersionRevisionDatesEntity {
+public abstract class RevisionDatesDmEntity extends RevisionDatesEntity {
 
     @Column(name = "deleted_at")
     protected LocalDateTime deletedAt;
@@ -69,8 +65,7 @@ public abstract class AbstractVersionRevisionDatesDeleteMarkerEntity extends Abs
 
     @Override
     public String toString() {
-        return "AbstractVersionRevisionDatesDeleteMarkerEntity [id=" + id + ", version=" + version + ", createdAt=" + createdAt + ", updatedAt=" + updatedAt
-                        + ", deletedAt=" + deletedAt + "]";
+        return "AbstractRevisionDatesDeleteMarkerEntity [id=" + id + ", createdAt=" + createdAt + ", updatedAt=" + updatedAt + ", deletedAt=" + deletedAt + "]";
     }
 
 }