+++ src/main/java/io/coodoo/framework/jpa/boundary/entity/dto/BaseOccEntityDTO.java
@@ -1,0 +1,21 @@
+package io.coodoo.framework.jpa.boundary.entity.dto;
+
+import io.coodoo.framework.jpa.boundary.entity.BaseOccEntity;
+
+/**
+ * Data transfer object (DTO) dedicated to {@link BaseOccEntity}
+ * 
+ * @author coodoo GmbH (coodoo.io)
+ */
+public class BaseOccEntityDTO extends BaseEntityDTO {
+
+    public Integer version;
+
+    public BaseOccEntityDTO() {}
+
+    public BaseOccEntityDTO(BaseOccEntity baseOccEntity) {
+        super(baseOccEntity);
+        this.version = baseOccEntity.getVersion();
+    }
+
+}