+++ src/main/java/io/coodoo/framework/jpa/boundary/entity/dto/RevisionDmOccEntityDTO.java
@@ -1,0 +1,21 @@
+package io.coodoo.framework.jpa.boundary.entity.dto;
+
+import io.coodoo.framework.jpa.boundary.entity.RevisionDmOccEntity;
+
+/**
+ * Data transfer object (DTO) dedicated to {@link RevisionDmOccEntity}
+ * 
+ * @author coodoo GmbH (coodoo.io)
+ */
+public class RevisionDmOccEntityDTO extends RevisionDmEntityDTO {
+
+    public Integer version;
+
+    public RevisionDmOccEntityDTO() {}
+
+    public RevisionDmOccEntityDTO(RevisionDmOccEntity revisionDmOccEntity) {
+        super(revisionDmOccEntity);
+        this.version = revisionDmOccEntity.getVersion();
+    }
+
+}