+++ src/main/java/io/coodoo/framework/jpa/boundary/entity/dto/RevisionEntityDTO.java
@@ -1,0 +1,23 @@
+package io.coodoo.framework.jpa.boundary.entity.dto;
+
+import io.coodoo.framework.jpa.boundary.entity.RevisionEntity;
+
+/**
+ * Data transfer object (DTO) dedicated to {@link RevisionEntity}
+ * 
+ * @author coodoo GmbH (coodoo.io)
+ */
+public class RevisionEntityDTO extends RevisionDatesEntityDTO {
+
+    public Long createdBy;
+    public Long updatedBy;
+
+    public RevisionEntityDTO() {}
+
+    public RevisionEntityDTO(RevisionEntity revisionEntity) {
+        super(revisionEntity);
+        this.createdBy = revisionEntity.getCreatedBy();
+        this.updatedBy = revisionEntity.getUpdatedBy();
+    }
+
+}